Improve view and test(2)
