+++ users/utils.py
@@ -1,9 +1,23 @@
 import jwt
 
+from datetime               import datetime, timedelta
+
 from django.conf import settings
 from django.http import JsonResponse
+from django.db.utils        import IntegrityError
+
 
 from users.models import CustomUser
+
+class Validation:
+    def check_duplicate_email(self, email):
+        if CustomUser.objects.filter(email = email).exists():
+            raise IntegrityError
+
+    def generate_jwt(self, user):
+        payload      = {'user_id': user.id, 'exp': datetime.now() +timedelta(hours=2)}
+        access_token = jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
+        return access_token
 
 def login_decorator(func):
     def wrapper(self,request,*args,**kwargs):
